created helper.c and .h and created seperate files for each algorithm
