+++ binary_search.c
@@ -1,0 +1,24 @@
+#include <stdio.h>
+#include <stdlib.h>
+
+int binary_search(int *arr, int key, int size){
+    int low = 0;
+    int high = size - 1;
+
+    while (low <= high)
+    {
+        int mid = (low + high) / 2;
+        if (arr[mid] == key)
+        {
+            return mid;
+        }
+        else if(arr[mid] < key){
+            low = mid + 1;
+        }
+        else{
+            high = mid - 1;
+        }
+    }
+    return 0;
+    
+}
