+++ helper.h
@@ -1,0 +1,7 @@
+#ifndef HELPER_H_
+#define HELPER_H_
+
+void randomize_array(int *arr, int size);
+int* array_malloc(int size);
+
+#endif
