+++ linear_binary_search.c
@@ -1,62 +1,7 @@
 #include <stdio.h>
 #include <stdlib.h>
+#include <helper.h>
 
-int* array_malloc(int size){
-    int *ptr = (int*) malloc(size*sizeof(int));
-
-    return ptr;
-}
-
-void randomize_array(int *arr, int size){
-    printf("Random Array: \n");
-    for (int i = 0; i < size; i++)
-    {
-        arr[i] = rand()% size; //wir brauchen ja eig kein random array fürs Suchen. (Binary funktioniert auch nicht mit random Array);
-        //arr[i] = i;  //zum Testen mit key=5 praktisch muss man nicht so lang debuggen
-        printf("%d, ", arr[i]);
-    }
-    printf("\n");
-}
-
-int linear_search(int *arr, int key, int size){
-    //int size = sizeof(arr) / sizeof(arr[0]); geht nicht weil sizeof(arr) die Größte des Pointers ist. Gibt keine Möglichkeit die Größe hier zu ermitteln oder? Kennst du was?
-
-    for (int i = 0; i < size; i++)
-    {
-        if (arr[i] == key)
-        {
-            printf("\n");
-            return i;
-        }
-    }
-    printf("\n");
-    return 0;
-
-}
-
-int binary_search(int *arr, int key, int size){
-    int low = 0;
-    int high = size - 1;
-
-    while (low <= high)
-    {
-        int mid = (low + high) / 2;
-        if (arr[mid] == key)
-        {
-            printf("\n");
-            return mid;
-        }
-        else if(arr[mid] < key){
-            low = mid + 1;
-        }
-        else{
-            high = mid - 1;
-        }
-    }
-    printf("\n");
-    return 0;
-    
-}
 
 int main(){
     int size = 100;
@@ -67,7 +12,7 @@ int main(){
     randomize_array(ptr, size);
 
     //ask for search algorithm
-    printf("Which algorithm to preform? (l=linear, b=binary): ");
+    printf("Which algorithm to perform? (l=linear, b=binary): ");
     char input;
     scanf(" %c", &input);
     printf("\n");
